Created seperate classes for each Component
